appointment_scheduler/requirements.txt
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
mongoengine==0.24.2